requirements.txt
apache-airflow==2.5.1
apache-airflow-providers-postgres==2.5.0
apache-airflow-providers-mongo==2.5.0
psycopg2==2.9.3
pymongo==3.12.1
requests==2.26.0
